* master:
  add dep annotation to rel scorecard
  add diff to rel score
  Add a grafana settings file without all the comments
  changes to speed up first impression
  eventDiff enhancements, SApp feedback
  cancel rel report query app vols by def
  protected from missing event locationd
  settings fix
  improve to event diff + fortify settings + improve to rel analysis

+++ src/main/java/com/takipi/integrations/grafana/settings/input/RegressionReportSettings.java
@@ -20,22 +20,22 @@ public class RegressionReportSettings {
 	/**
 	 *  The number of points deducted for a new event 
 	 */
-	public int new_event_score;
+	public double new_event_score;
 	
 	/**
 	 *  The number of points deducted for a regressed event
 	 */
-	public int regression_score;
+	public double regression_score;
 	
 	/**
 	 *  The number of points deducted for a severe new event 
 	 */
-	public int severe_new_event_score;
+	public double severe_new_event_score;
 	
 	/**
 	 *  The number of points deducted for a severe regression or slowdown
 	 */
-	public int critical_regression_score;
+	public double critical_regression_score;
 	
 	/**
 	 * A factor applied to the score deduction of a new error, regression or slowdown.
